Create objects in source in the structure they're expected at the dest
We are changing our S3 sync jobs to do a straight copy of objects from
source to destination bucket. This will make it much easier for us to
maintain the sync jobs, and the sync jobs will be much more fault
tolerant as well.

Because of this change, we can no longer transform the subscription
egress data at sync time. This change updates the export command/job to
place the source objects into the folder structure that the downstream
data consumers expect.